refactor: modulize getComputerNumber
